Add tile Id to interfaces

+++ source/renderers/VexTile.Renderer.Common/Interfaces/ISymbol.cs
@@ -2,5 +2,5 @@
 
 public interface ISymbol
 {
-    uint TileId { get; }
+    ulong TileId { get; }
 }
